+++ CHANGES.rst
@@ -21,10 +21,15 @@ Unreleased
 - Configuration values which are file paths will now apply tilde-expansion,
   closing `issue 589`_.
 
+- Now secondary config files like tox.ini and setup.cfg can be specified
+  explicitly, and prefixed sections like `[coverage:run]` will be read. Fixes
+  `issue 588`_.
+
 - Be more flexible about the command name displayed by help, fixing
   `issue 600`_. Thanks, Ben Finney.
 
 .. _issue 581: https://bitbucket.org/ned/coveragepy/issues/581/race-condition-when-saving-data-under
+.. _issue 588: https://bitbucket.org/ned/coveragepy/issues/588/using-rcfile-path-to-toxini-uses-run
 .. _issue 589: https://bitbucket.org/ned/coveragepy/issues/589/allow-expansion-in-coveragerc
 .. _issue 600: https://bitbucket.org/ned/coveragepy/issues/600/get-program-name-from-command-line-when